+++ compiler/src/jdk.graal.compiler/src/jdk/graal/compiler/lir/gen/LIRGenerator.java
@@ -531,7 +531,7 @@ public abstract class LIRGenerator extends CoreProvidersDelegate implements LIRG
      * </ol>
      */
     public void emitStrategySwitch(JavaConstant[] keyConstants, double[] keyProbabilities, LabelRef[] keyTargets, LabelRef defaultTarget, AllocatableValue value,
-                    boolean inputMayBeOutOfRange) {
+                    boolean inputMayBeOutOfRange, boolean mayEmitThreadedCode) {
         SwitchStrategy strategy = SwitchStrategy.getBestStrategy(keyProbabilities, keyConstants, keyTargets);
 
         if (strategy.getAverageEffort() < JUMP_TABLE_THRESHOLD) {
@@ -542,7 +542,7 @@ public abstract class LIRGenerator extends CoreProvidersDelegate implements LIRG
         // If the density of the jump table would be larger than this value then we will emit one
         double minDensity = 1 / Math.sqrt(strategy.getAverageEffort());
 
-        Subrange subrangeForDirectJump = findSubrangeForDirectJumpTable(keyConstants, keyProbabilities, minDensity);
+        Subrange subrangeForDirectJump = findSubrangeForDirectJumpTable(keyConstants, keyProbabilities, minDensity, mayEmitThreadedCode);
         if (subrangeForDirectJump != null) {
             int loIdx = subrangeForDirectJump.loIdx;
 
@@ -555,13 +555,13 @@ public abstract class LIRGenerator extends CoreProvidersDelegate implements LIRG
                         break;
                     }
                 }
-                if (directJumpForSubrange(keyConstants, minDensity, 0, extendedLoIdx, subrangeForDirectJump.hiIdx)) {
+                if (mayEmitThreadedCode || directJumpForSubrange(keyConstants, minDensity, 0, extendedLoIdx, subrangeForDirectJump.hiIdx)) {
                     loIdx = extendedLoIdx;
                     minValue = 0;
                 }
             }
 
-            emitDirectJumpTableHelper(keyConstants, keyProbabilities, keyTargets, defaultTarget, value, minValue, loIdx, subrangeForDirectJump.hiIdx, inputMayBeOutOfRange);
+            emitDirectJumpTableHelper(keyConstants, keyProbabilities, keyTargets, defaultTarget, value, minValue, loIdx, subrangeForDirectJump.hiIdx, inputMayBeOutOfRange, mayEmitThreadedCode);
             return;
         }
 
@@ -600,8 +600,8 @@ public abstract class LIRGenerator extends CoreProvidersDelegate implements LIRG
      * jump table with them, the default target will now check the other remaining targets. In the
      * best case, our subrange will cover the whole range and there is no remaining target.
      */
-    private static Subrange findSubrangeForDirectJumpTable(JavaConstant[] keyConstants, double[] keyProbabilities, double minDensity) {
-        if (directJumpForSubrange(keyConstants, minDensity, keyConstants[0].asInt(), 0, keyConstants.length - 1)) {
+    private static Subrange findSubrangeForDirectJumpTable(JavaConstant[] keyConstants, double[] keyProbabilities, double minDensity, boolean mayEmitThreadedCode) {
+        if (mayEmitThreadedCode || directJumpForSubrange(keyConstants, minDensity, keyConstants[0].asInt(), 0, keyConstants.length - 1)) {
             // If all cases can be fit into a jump table then ignore the probability
             return new Subrange(0, keyConstants.length - 1);
         }
@@ -641,7 +641,7 @@ public abstract class LIRGenerator extends CoreProvidersDelegate implements LIRG
     }
 
     private void emitDirectJumpTableHelper(JavaConstant[] keyConstants, double[] keyProbs, LabelRef[] keyTargets, LabelRef defaultTarget, AllocatableValue value, int minValue, int loIdx, int hiIdx,
-                    boolean inputMayBeOutOfRange) {
+                    boolean inputMayBeOutOfRange, boolean mayEmitThreadedCode) {
         int maxValue = keyConstants[hiIdx].asInt();
         // This cannot overflow because we have ensured that above
         int subrangeValueRange = maxValue - minValue + 1;
@@ -681,7 +681,7 @@ public abstract class LIRGenerator extends CoreProvidersDelegate implements LIRG
                 jumpTableEntry.getTargetBlock().setAlign(true);
             }
         }
-        emitRangeTableSwitch(minValue, defaultTarget, targets, remainingStrategy, remainingKeyTargets, value, inputMayBeOutOfRange);
+        emitRangeTableSwitch(minValue, defaultTarget, targets, remainingStrategy, remainingKeyTargets, value, inputMayBeOutOfRange, mayEmitThreadedCode);
     }
 
     private void emitHashedJumpTableHelper(JavaConstant[] keyConstants, LabelRef[] keyTargets, LabelRef defaultTarget, AllocatableValue value, IntHasher h) {
@@ -706,8 +706,9 @@ public abstract class LIRGenerator extends CoreProvidersDelegate implements LIRG
     private static final int MAX_DIRECT_SIZE = 128;
 
     /**
-     * Each label is 4 bytes in the direct jump table, so if the table is too sparse we may waste a
-     * lot of space. As a result, we emit a table if it is not too large, OR if it is dense enough.
+     * Each jump table entry is 4 bytes on AArch64 or 8 bytes on AMD64, so if the jump table is too
+     * sparse we may waste a lot of space. As a result, we emit a table if it is not too large, OR
+     * if it is dense enough.
      */
     private static boolean directJumpForSubrange(JavaConstant[] keyConstants, double minDensity, int minValue, int loIdx, int hiIdx) {
         // Subtraction of 2 sorted signed values will not overflow the unsigned range
@@ -731,7 +732,7 @@ public abstract class LIRGenerator extends CoreProvidersDelegate implements LIRG
      * defaultTarget} when the value of {@code key} is not in the jump table.
      */
     protected abstract void emitRangeTableSwitch(int lowKey, LabelRef defaultTarget, LabelRef[] targets, SwitchStrategy remainingStrategy, LabelRef[] remainingTargets, AllocatableValue key,
-                    boolean inputMayBeOutOfRange);
+                    boolean inputMayBeOutOfRange, boolean mayEmitThreadedCode);
 
     protected abstract void emitHashTableSwitch(JavaConstant[] keys, LabelRef defaultTarget, LabelRef[] targets, AllocatableValue value, Value hash);